a function that reads a text file and prints it to the POSIX standard output

+++ 0x15-file_io/0-read_textfile.c
@@ -10,7 +10,7 @@
 
 ssize_t read_textfile(const char *filename, size_t letters)
 {
-ssize_t file, i, n;
+int file, i, n;
 char *buf;
 
 if (!filename)
